+++ script/github-actions/daily-product-scan/main.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 const glob = require('glob');
 const core = require('@actions/core');
+const fs = require('fs');
 
 const Products = require('./products');
 const PackageDependencies = require('./products/dependencies/package-dependencies');
@@ -69,6 +70,27 @@ async function main({ octokit }) {
 
   const productDirectory = JSON.parse(response.data);
 
+  // Check for any products that have been added and are not present in GitHub, then add them to the directory.
+
+  const manifestIds = Object.keys(products.all);
+  const productListIds = productDirectory.map(product => product.product_id);
+
+  manifestIds.forEach(id => {
+    if (id.length === 36 && productListIds.indexOf(id) === -1) {
+      const manifest = JSON.parse(
+        fs.readFileSync(`${products.all[id].pathToCode}/manifest.json`),
+      );
+      const product = {
+        // eslint-disable-next-line camelcase
+        product_id: manifest.productId,
+        // eslint-disable-next-line camelcase
+        product_name: manifest.appName,
+      };
+      productDirectory.push(product);
+    }
+  });
+
+  // Check for automatically updated field values
   const differ = new Differ();
   const updatedProductDirectory = differ.diff({
     products,
